SVCPLAN-2588: add tempsetDNS postscript
The tempsetDNS postscript is used to configure NCSA DNS in /etc/resolv.conf. This
is needed after running confignetwork with the -s flag to set interfaces to static.
When this is run Network Manager clobbers /etc/resolv.conf and we've been unable
to figure out how to get xCAT to avoid/correct this "natively". So this would be
run after 'confignetwork -s' and prior to anything that requires functioning DNS,
e.g., Puppet.

Also adds execute perms on postscripts/simple_fw.